requirements.txt
# Core dependencies
google-genai>=1.60.0
python-dotenv>=1.0.1
Pillow>=12.1.0
pydantic>=2.12.5

# API server
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.1

# Layer extraction (fal.ai/Qwen)
fal-client>=0.12.0
aiofiles>=23.2.1
